Move can.Stub to the separate file
