+++ src/ggrc-client/js/models/index.js
@@ -5,6 +5,7 @@
 
 import './mappers';
 import './save_queue';
+import './stub';
 import './models-extensions';
 
 import './business-models';
